Polybar : remove buggy tray for now
The actual solution is probably to use what patrick96 proposed on
irc.freenode.net#polybar : disable the tray in polybar and create a mini
tray using tint2. The tray_tint2rc file in the tint2 folder has a basic
configuration for the tray, but I don't manage yet the width of the
polybar to make room for the tray.

Note : in this solution all sizes are fixed so the count of icons in the
tray has to be kept minimal