Zsh : fix completion
The completion functions have their own folder, and the fpath must be up
to date (or at least hold all the completions functions) when the call
to compinit is made, since it automagically crawls fpath to find
completion definitions.

[Reference](http://zsh.sourceforge.net/Doc/Release/Completion-System.html#Use-of-compinit)

The doc is still a pain to read though, so I am not 100% positive.

Also, Cargo folder under plugin has been entirely removed since it only
held a completion function